Change approach to sending labels to OpenAI
Generate a numbered list of labels, in the order they're found in the form JSON, so that we don't need to send the overhead of the JSON encoding.
Ask OpenAI to send back a numbered list of unique IDs.
Store the paths to each label, and then use the path to set the new ID.